global effects provider 2/5: effect preview in execution scenario
