+++ src/effects/builtin/view/builtineffectmodel.cpp
@@ -59,7 +59,7 @@ void BuiltinEffectModel::doStartPreview()
 
 void BuiltinEffectModel::doStopPreview()
 {
-    effectsProvider()->stopPreview();
+    executionScenario()->stopPreview();
 }
 
 void BuiltinEffectModel::modifySettings(const std::function<void(EffectSettings& settings)>& modifier)
