+++ src/effects/effects_base/ieffectexecutionscenario.h
@@ -25,5 +25,6 @@ public:
     virtual muse::Ret repeatLastProcessor() = 0;
 
     virtual muse::Ret previewEffect(const EffectInstanceId& effectInstanceId, EffectSettings& settings) = 0;
+    virtual void stopPreview() = 0;
 };
 }
