+++ src/effects/effects_base/internal/effectexecutionscenario.cpp
@@ -3,12 +3,16 @@
 */
 #include "effectexecutionscenario.h"
 
+#include "framework/global/async/async.h"
 #include "framework/global/defer.h"
 #include "framework/global/realfn.h"
 #include "framework/global/translation.h"
 
+#include "playback/iplayer.h"
+
 #include "au3-project/Project.h"
 #include "au3-effects/Effect.h"
+#include "au3-effects/MixAndRender.h"
 #include "au3-module-manager/PluginManager.h"
 
 #include "au3-track/Track.h"
@@ -18,7 +22,9 @@
 #include "au3-effects/EffectManager.h"
 #include "au3-module-manager/ConfigInterface.h"
 #include "au3-numeric-formats/NumericConverterFormats.h"
+#include "au3-numeric-formats/ProjectTimeSignature.h"
 #include "au3-transactions/TransactionScope.h"
+#include "au3-stretching-sequence/TempoChange.h"
 
 #include "au3wrap/au3types.h"
 #include "au3wrap/internal/domaccessor.h"
@@ -485,7 +491,7 @@ muse::async::Channel<EffectId> EffectExecutionScenario::lastProcessorIdChanged()
 muse::Ret EffectExecutionScenario::previewEffect(const EffectInstanceId& effectInstanceId, EffectSettings& settings)
 {
     EffectId effectId = effectInstancesRegister()->effectIdByInstanceId(effectInstanceId);
-    return effectsProvider()->previewEffect(effectId, settings);
+    return doPreviewEffect(effectId, settings);
 }
 
 namespace {
@@ -616,4 +622,253 @@ muse::Ret EffectExecutionScenario::performEffectInternal(au3::Au3Project& projec
     }
 
     return success;
+}
+
+namespace {
+void restoreEffectStateHack(EffectBase& effect)
+{
+    if (auto pInstance = std::dynamic_pointer_cast<EffectInstanceEx>(effect.MakeInstance())) {
+        pInstance->Init();
+    }
+}
+}
+
+muse::Ret EffectExecutionScenario::doPreviewEffect(const EffectId& effectId, EffectSettings& settings)
+{
+    ::EffectBase* pEffect = effectsProvider()->effect(effectId);
+    if (!pEffect) {
+        return muse::make_ret(muse::Ret::Code::InternalError);
+    }
+    auto& effect = *pEffect;
+
+    const bool isNyquist = effect.GetFamily() == NYQUISTEFFECTS_FAMILY;
+    const bool isGenerator = effect.GetType() == EffectTypeGenerate;
+
+    //! ============================================================================
+    //! NOTE Step 1 - check conditions
+    //! ============================================================================
+    if (effect.mNumTracks == 0 && !isGenerator) {     // nothing to preview
+        return muse::make_ret(muse::Ret::Code::InternalError);
+    }
+
+    auto player = playback()->player();
+    if (player->isBusy()) {
+        LOGW() << "can't play, maybe audio is busy";
+        return muse::make_ret(muse::Ret::Code::InternalError);
+    }
+
+    //! ============================================================================
+    //! NOTE Step 2 - save origin context (state)
+    //! ============================================================================
+    const EffectContext originCtx = { effect.mT0, effect.mT1, effect.mTracks, effect.mProgress, effect.mIsPreview };
+
+    // restore internal effect state on return (if needed)
+    auto cleanup0 = effect.BeginPreview(settings);
+
+    // Effect is already inited; we will call Process and then Init
+    // again, so the state is exactly the way it was before Preview
+    // was called.
+    auto cleanup1 = finally([&] {
+        restoreEffectStateHack(effect);
+    });
+
+    //! ============================================================================
+    //! NOTE Step 3 - make new context (state)
+    //! ============================================================================
+
+    EffectContext newCtx;
+
+    //! Step 3.1 - prepare time
+    newCtx.t0 = originCtx.t0;
+    if (effect.PreviewsFullSelection()) {
+        newCtx.t1 = originCtx.t1;
+    } else {
+        // Limit preview time:
+        // We need to pre-render the audio, which would take a long time and lots of memory for long selections.
+        // On the other hand, preview isn't typically something users would listen to for more than a few seconds.
+        // (Au3 used to read `previewLen` from the `/AudioIO/EffectsPreviewLen` setting.
+        // There is no plan at the moment to reintroduce it in Au4.)
+        const double maxPreviewLen = effectsConfiguration()->previewMaxDuration();
+        const double previewLen = std::min(originCtx.t1 - originCtx.t0, maxPreviewLen);
+        double previewDuration = 0.0;
+        if (isNyquist && isGenerator) {
+            previewDuration = effect.CalcPreviewInputLength(settings, previewLen);
+        } else {
+            previewDuration = std::min(settings.extra.GetDuration(), effect.CalcPreviewInputLength(settings, previewLen));
+        }
+        newCtx.t1 = originCtx.t0 + previewDuration;
+    }
+
+    if ((newCtx.t1 > originCtx.t1) && !isGenerator) {
+        newCtx.t1 = originCtx.t1;
+    }
+
+    if (muse::RealIsEqualOrLess(newCtx.t1, newCtx.t0)) {
+        return muse::make_ret(muse::Ret::Code::InternalError);
+    }
+
+    //! Step 3.2 - make new tracks
+    {
+        // Build NEW tracklist from rendering tracks
+        // Set the same owning project, so FindProject() can see it within Process()
+        const auto pProject = effect.mTracks->GetOwner();
+        newCtx.tracks = au::au3::Au3TrackList::Create(pProject);
+
+        // Linear Effect preview optimised by pre-mixing to one track.
+        // Generators need to generate per track.
+        const bool isLinearEffect = effect.IsLinearEffect();
+        if (isLinearEffect && !isGenerator) {
+            auto newTrack = MixAndRender(
+                originCtx.tracks->Selected<const au::au3::Au3WaveTrack>(),
+                Mixer::WarpOptions { pProject },
+                wxString {}, // Don't care about the name of the temporary tracks
+                effect.mFactory, effect.mProjectRate, floatSample, newCtx.t0, newCtx.t1);
+
+            if (!newTrack) {
+                return muse::make_ret(muse::Ret::Code::InternalError);
+            }
+
+            newCtx.tracks->Add(newTrack);
+
+            newTrack->MoveTo(0);
+            newTrack->SetSelected(true);
+            std::static_pointer_cast<WaveTrack>(newTrack)->SetMute(false);
+        } else {
+            if (effect.mNumTracks > 0) {
+                for (const WaveTrack* src : originCtx.tracks->Selected<const au::au3::Au3WaveTrack>()) {
+                    const auto dest = std::static_pointer_cast<WaveTrack>(src->Copy(newCtx.t0, newCtx.t1));
+                    dest->SetSelected(true);
+                    dest->SetMute(false);
+                    newCtx.tracks->Add(dest);
+                }
+            } else {
+                // If we are previewing a generator without a given track, we can just give mono.
+                constexpr auto nChannels = 1;
+                const auto dest = WaveTrackFactory::Get(*pProject).Create(nChannels);
+                dest->SetSelected(true);
+                const auto tempo = ProjectTimeSignature::Get(*pProject).GetTempo();
+                DoProjectTempoChange(*dest, tempo);
+                newCtx.tracks->Add(dest);
+            }
+        }
+    }
+
+    // NEW tracks start at time zero.
+    // Adjust T0 and T1 to be the times to process, and to
+    // play back in these tracks
+    double startOffset = newCtx.t0;
+    newCtx.t1 -= newCtx.t0;
+    newCtx.t0 = 0.0;
+
+    //! ============================================================================
+    //! NOTE Step 4 - process
+    //! ============================================================================
+    {
+        using namespace BasicUI;
+        auto progress = MakeProgress(
+            effect.GetName(),
+            XO("Preparing preview"),
+            ProgressShowStop
+            ); // Have only "Stop" button.
+
+        newCtx.preparingPreviewProgress = progress.get();
+        newCtx.isPreview = true;
+
+        // apply new context
+        {
+            effect.mT0 = newCtx.t0;
+            effect.mT1 = newCtx.t1;
+            effect.mTracks = newCtx.tracks;
+            effect.mProgress = newCtx.preparingPreviewProgress;
+            effect.mIsPreview = newCtx.isPreview;
+
+            // Update track/group counts
+            effect.CountWaveTracks();
+        }
+
+        // Apply effect to new tracks
+
+        auto pInstance = std::dynamic_pointer_cast<EffectInstanceEx>(effect.MakeInstance());
+        IF_ASSERT_FAILED(pInstance) {
+            return muse::make_ret(muse::Ret::Code::InternalError);
+        }
+
+        bool success = pInstance->Process(settings);
+        if (!success) {
+            return muse::make_ret(muse::Ret::Code::InternalError);
+        }
+    }
+
+    //! ============================================================================
+    //! NOTE Step 4 - play new processed tracks
+    //! ============================================================================
+    {
+        playback::PlayTracksOptions opt;
+        opt.selectedOnly = true;
+        opt.startOffset = startOffset;
+        opt.isDefaultPolicy = false;
+
+        // Setting looping to `false` ensures that the loop region won't interfere with preview.
+        // Also, looping is of course not effective during preview. Having it visually disabled
+        // makes it clear to the user.
+        const auto loopWasActive = player->isLoopRegionActive();
+        player->setLoopRegionActive(false);
+        player->setPlaybackRegion({ startOffset + newCtx.t0, startOffset + newCtx.t1 });
+
+        m_effectPreviewState.emplace(effectId, originCtx, newCtx.tracks, loopWasActive);
+
+        // Listen for playback stop to clean up preview state when audio finishes
+        // before user manually stops it.
+        player->playbackStatusChanged().onReceive(this, [this](playback::PlaybackStatus status) {
+            if (status == playback::PlaybackStatus::Stopped) {
+                // Wait for other observers of this signal to be finished, in particular the audio engine,
+                // which should stop playback synchronously. Only then we may delete the preview tracks.
+                async::Async::call(this, [this] {
+                    stopPreview();
+                });
+            }
+        });
+
+        muse::Ret ret = player->playTracks(*newCtx.tracks, newCtx.t0, newCtx.t1, opt);
+        if (!ret) {
+            player->playbackStatusChanged().disconnect(this);
+            m_effectPreviewState.reset();
+            return ret;
+        }
+    }
+
+    return muse::make_ok();
+}
+
+void EffectExecutionScenario::stopPreview()
+{
+    if (!m_effectPreviewState) {
+        return;
+    }
+
+    const auto player = playback()->player();
+
+    IF_ASSERT_FAILED(player) {
+        return;
+    }
+
+    player->playbackStatusChanged().disconnect(this);
+
+    if (player->playbackStatus() == playback::PlaybackStatus::Running) {
+        player->stop();
+    }
+
+    EffectBase* effect = effectsProvider()->effect(m_effectPreviewState->effectId);
+    if (effect) {
+        const EffectContext& originCtx = m_effectPreviewState->originContext;
+        effect->mT0 = originCtx.t0;
+        effect->mT1 = originCtx.t1;
+        effect->mTracks = originCtx.tracks;
+        effect->mProgress = originCtx.preparingPreviewProgress;
+        effect->mIsPreview = originCtx.isPreview;
+    }
+
+    player->setLoopRegionActive(m_effectPreviewState->loopWasActive);
+
+    m_effectPreviewState.reset();
 }
