+++ src/effects/vst/view/vstviewmodel.h
@@ -11,7 +11,6 @@
 #include "effects/effects_base/irealtimeeffectservice.h"
 #include "effects/effects_base/effectstypes.h"
 #include "effects/effects_base/view/abstracteffectviewmodel.h"
-#include "effects/effects_base/ieffectsprovider.h"
 
 class VST3Instance;
 class EffectSettingsAccess;
@@ -23,7 +22,6 @@ class VstViewModel : public AbstractEffectViewModel
 public:
     muse::ContextInject<IRealtimeEffectService> realtimeEffectService{ this };
     muse::ContextInject<trackedit::IProjectHistory> projectHistory{ this };
-    muse::ContextInject<IEffectsProvider> effectsProvider{ this };
 
 public:
     VstViewModel(QObject* parent, int instanceId);
